requirements.txt
asgiref==3.2.10
Babel==2.8.0
boto3==1.16.35
botocore==1.19.35
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-crispy-forms==1.9.2
django-easy-pdf==0.1.1
django-google-maps==0.12.2
django-import-export==2.3.0
django-phone-field==1.8.1
django-phone-verify==0.1.1
django-phonenumber-field==4.0.0
django-storages==1.10.1
djangorestframework==3.11.1
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
phonenumbers==8.12.7
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.48
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
tinycss2==1.0.2
twilio==6.44.2
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
